Move encoding-support to subdir, adjust includes and Makefiles
