+++ compiler/make/Makefile.compiler.head
@@ -35,7 +35,9 @@ endif
 
 include $(CHPL_MAKE_HOME)/make/Makefile.base
 
-INCL_CFLAGS = -I. -I$(COMPILER_ROOT)/include/$(CHPL_MAKE_HOST_PLATFORM) -I$(COMPILER_ROOT)/include $(LLVM_INCLUDES) -I$(COMPILER_BUILD) -I$(THIRD_PARTY_DIR)/utf8-decoder
+RUNTIME_ROOT = $(COMPILER_ROOT)/../runtime
+
+INCL_CFLAGS = -I. -I$(COMPILER_ROOT)/include/$(CHPL_MAKE_HOST_PLATFORM) -I$(COMPILER_ROOT)/include $(LLVM_INCLUDES) -I$(COMPILER_BUILD) -I$(RUNTIME_ROOT)/include/encoding -I$(THIRD_PARTY_DIR)/utf8-decoder
 COMP_CXXFLAGS += $(INCL_CFLAGS)
 
 #
